OCR/requirements.txt
fastapi
uvicorn[standard]
jinja2
pytest
requests
aiofiles
python-multipart
pillow
pytesseract